+++ src/api/GoalsApi.js
@@ -14,6 +14,7 @@
 
 import ApiClient from "../ApiClient";
 import AddOrUpdateGoalResponse200 from '../model/AddOrUpdateGoalResponse200';
+import BasicGoal from '../model/BasicGoal';
 import DeleteGoalResponse200 from '../model/DeleteGoalResponse200';
 import GetGoalResultResponse200 from '../model/GetGoalResultResponse200';
 import GetGoalsResponse200 from '../model/GetGoalsResponse200';
@@ -367,27 +368,17 @@ export default class GoalsApi {
      * Updates an existing goal.
      * @param {String} id The ID of the goal to be updated
      * @param {Object} opts Optional parameters
-     * @param {String} opts.title The title of the goal
-     * @param {Object.<String, Object>} opts.assignee Who this goal is assigned to. It requires the following JSON structure: { \\\"id\\\": \\\"1\\\", \\\"type\\\": \\\"person\\\" }. `type` can be either `person`, `company` or `team`. ID of the assignee person, company or team.
-     * @param {Object.<String, Object>} opts.type The type of the goal. It requires the following JSON structure: { \\\"name\\\": \\\"deals_started\\\", \\\"params\\\": { \\\"pipeline_id\\\": 1 } }. Type can be one of: `deals_won`, `deals_progressed`, `activities_completed`, `activities_added`, `deals_started` or `revenue_forecast`. `params` can include `pipeline_id`, `stage_id` or `activity_type_id`. `stage_id` is related to only `deals_progressed` type of goals and `activity_type_id` to `activities_completed` or `activities_added` types of goals. To track goal in all pipelines set `pipeline_id` as `null`.
-     * @param {Object.<String, Object>} opts.expectedOutcome The expected outcome of the goal. Expected outcome can be tracked either by `quantity` or by `sum`. It requires the following JSON structure: { \\\"target\\\": \\\"50\\\", \\\"tracking_metric\\\": \\\"quantity\\\" } or { \\\"target\\\": \\\"50\\\", \\\"tracking_metric\\\": \\\"sum\\\", \\\"currency_id\\\": 1 }. `currency_id` should only be added to `sum` type of goals.
-     * @param {Object.<String, Object>} opts.duration The date when the goal starts and ends. It requires the following JSON structure: { \\\"start\\\": \\\"2019-01-01\\\", \\\"end\\\": \\\"2022-12-31\\\" }. Date in format of YYYY-MM-DD. \\\"end\\\" can be set to `null` for an infinite, open-ended goal.
-     * @param {module:model/String} opts.interval The interval of the goal
+     * @param {module:model/BasicGoal} opts.basicGoal 
      * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with an object containing data of type {@link module:model/AddOrUpdateGoalResponse200} and HTTP response
      */
     updateGoalWithHttpInfo(id, opts) {
       opts = opts || {};
-      let postBody = null;
+      let postBody = opts['basicGoal'];
 
       // verify the required parameter 'id' is set
       if (id === undefined || id === null) {
         throw new Error("Missing the required parameter 'id' when calling updateGoal");
       }
-
-
-
-
-
 
 
       let pathParams = {
@@ -398,24 +389,12 @@ export default class GoalsApi {
       let headerParams = {
       };
       let formParams = {
-        'title': opts['title'],
-        'assignee': opts['assignee'],
-        'type': opts['type'],
-        'expected_outcome': opts['expectedOutcome'],
-        'duration': opts['duration'],
-        'interval': opts['interval'],
       };
 
       let formParamArray = [
-        'title',
-        'assignee',
-        'type',
-        'expectedOutcome',
-        'duration',
-        'interval',
       ];
 
-      let contentTypes = ['application/x-www-form-urlencoded', ];
+      let contentTypes = ['application/json', ];
       const isURLEncoded = contentTypes.includes('application/x-www-form-urlencoded');
       const isJSON = contentTypes.includes('application/json');
 
@@ -444,12 +423,7 @@ export default class GoalsApi {
      * Updates an existing goal.
      * @param {String} id The ID of the goal to be updated
      * @param {Object} opts Optional parameters
-     * @param {String} opts.title The title of the goal
-     * @param {Object.<String, Object>} opts.assignee Who this goal is assigned to. It requires the following JSON structure: { \\\"id\\\": \\\"1\\\", \\\"type\\\": \\\"person\\\" }. `type` can be either `person`, `company` or `team`. ID of the assignee person, company or team.
-     * @param {Object.<String, Object>} opts.type The type of the goal. It requires the following JSON structure: { \\\"name\\\": \\\"deals_started\\\", \\\"params\\\": { \\\"pipeline_id\\\": 1 } }. Type can be one of: `deals_won`, `deals_progressed`, `activities_completed`, `activities_added`, `deals_started` or `revenue_forecast`. `params` can include `pipeline_id`, `stage_id` or `activity_type_id`. `stage_id` is related to only `deals_progressed` type of goals and `activity_type_id` to `activities_completed` or `activities_added` types of goals. To track goal in all pipelines set `pipeline_id` as `null`.
-     * @param {Object.<String, Object>} opts.expectedOutcome The expected outcome of the goal. Expected outcome can be tracked either by `quantity` or by `sum`. It requires the following JSON structure: { \\\"target\\\": \\\"50\\\", \\\"tracking_metric\\\": \\\"quantity\\\" } or { \\\"target\\\": \\\"50\\\", \\\"tracking_metric\\\": \\\"sum\\\", \\\"currency_id\\\": 1 }. `currency_id` should only be added to `sum` type of goals.
-     * @param {Object.<String, Object>} opts.duration The date when the goal starts and ends. It requires the following JSON structure: { \\\"start\\\": \\\"2019-01-01\\\", \\\"end\\\": \\\"2022-12-31\\\" }. Date in format of YYYY-MM-DD. \\\"end\\\" can be set to `null` for an infinite, open-ended goal.
-     * @param {module:model/String} opts.interval The interval of the goal
+     * @param {module:model/BasicGoal} opts.basicGoal 
      * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with data of type {@link module:model/AddOrUpdateGoalResponse200}
      */
     updateGoal(id, opts) {